Finishin part4. Exercises until 4.23.

+++ part4/bloglist/src/controllers/blogRoutes.js
@@ -1,32 +1,47 @@
 const route = require('express').Router()
 const Blog = require('../models/blogModel')
-const User = require('../models/userModel')
+const UserModel = require('../models/userModel')
+const { userExtractor } = require('../utils/middleware')
 
 route.get('/', async (request, response) => {
+  
   const blogs = await Blog.find({}).populate('user', {
     blogs: false
   })
-  
   response.json(blogs)
 })
 
-route.post('/', async (request, response, next) => {
-  const user = await User.findOne({})
+
+route.post('/', userExtractor, async (request, response) => {
+  const userData = request.user
+  
   const blog = new Blog({
     ...request.body,
-    user: user.id
+    user: userData.id
   })
-  const result = await blog.save()
-  user.blogs = user.blogs.concat(result.id)
-  await user.save()
-  response.status(201).json(result)
+  const persistedBlog = await blog.save()
+
+  await UserModel.updateOne(
+    { _id: userData.id},
+    { $push: {blogs: persistedBlog.id}}
+  )
+
+  response.status(201).json(persistedBlog)
 })
 
-route.delete('/:id', async (request, response) => {
-  const del = await Blog.findByIdAndDelete(request.params.id)
-  if(del)
-    response.status(204).json({message: 'OK'})
-  response.status(404).end()
+route.delete('/:id', userExtractor, async (request, response) => {
+  const userData = request.user
+  const blogToDelete = await Blog.findById(request.params.id)
+  if(!blogToDelete){
+    response.status(404).end()
+  }
+  if(blogToDelete.user && blogToDelete.user.toString() !== userData.id.toString()){
+    return response.status(403).json({message: 'Only the owner of the blog can deleted!'})
+  }
+  
+  await Blog.findByIdAndDelete(request.params.id)
+  response.status(204).json({message: 'OK'})
+  
 })
 
 route.put('/:id', async (request, response) => {
